+++ src/getScreenSize/getScreenSize.stories.js
@@ -33,7 +33,12 @@ function LocalResponsive() {
   return (
     <div
       ref={element}
-      style={{ height: '50vh', width: '50vw', backgroundColor: '#eee' }}
+      style={{
+        height: '50vh',
+        width: '50vw',
+        backgroundColor: '#eee',
+        overflow: 'auto',
+      }}
     >
       <p>Resize me</p>